+++ packages/manifest/src/StatsManager.ts
@@ -84,7 +84,7 @@ class StatsManager {
 
       assert(remoteEntryNameChunk, 'Can not get remoteEntry chunk!');
       const files = Array.from(remoteEntryNameChunk.files).filter(
-        (f) => !f.includes(HOT_UPDATE_SUFFIX),
+        (f) => !f.includes(HOT_UPDATE_SUFFIX) && !f.endsWith('.css'),
       );
       assert(
         files.length === 1,
@@ -440,9 +440,9 @@ class StatsManager {
       return false;
     } else if (publicPath === 'auto') {
       console.warn(
-        chalk`{bold {yellow [ ${PLUGIN_IDENTIFIER} ]: Manifest will not generate, because publicPath can only be absolute path, but got '${publicPath}'}}`,
+        chalk`{bold {blue [ ${PLUGIN_IDENTIFIER} ]: Manifest will use absolute path resolution via its host at runtime, reason: publicPath='${publicPath}'}}`,
       );
-      return false;
+      return true;
     }
 
     return true;