+++ packages/runtime/src/utils/plugin.ts
@@ -8,6 +8,8 @@ export function registerPlugins(
   hookInstances: Array<
     | FederationHost['hooks']
     | FederationHost['snapshotHandler']['hooks']
+    | FederationHost['sharedHandler']['hooks']
+    | FederationHost['remoteHandler']['hooks']
     | Module['host']['loaderHook']
   >,
 ) {