+++ apps/runtime-demo/3008-runtime-remote/CHANGELOG.md
@@ -1,5 +1,12 @@
 # 3008-runtime-remote
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies [0113b81]
+  - @module-federation/enhanced@0.1.13
+
 ## 1.0.1
 
 ### Patch Changes